
\name{stable_mle_fit}
\alias{stable_mle_fit}
\title{Fit a stable distribution to a sample using maximum likelihood}
\usage{
stable_mle_fit(x, init_vals, trace)
}
\arguments{
    \item{x}{sample vector}
    \item{init_vals}{initial guess for parameters. Defaults to NULL in which case these are set to defaults}
    \item{trace}{trace level}
}
\description{
    Fit a stable distribution to a sample using maximum likelihood
}
